requirements.txt
python-dotenv
motor